functions/requirements.txt
google-cloud-storage>=2
google-cloud-firestore>=2
ffmpeg-python>=0.2
tqdm>=4
